requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
mysqlclient==2.0.3
PyJWT==2.0.1
pytz==2020.5
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.2
